Moved types to @/lib/types.ts
