+++ app/context/courseContext.tsx
@@ -18,13 +18,6 @@ const CourseContext = createContext<CourseData>({
   setCurrentCourse: (): string => "",
 });
 
-export type Course = {
-  id: number;
-  name: string;
-};
-
-export type CourseList = Course[];
-
 export const CourseContextProvider = ({
   children,
 }: {
