+++ lib/types.ts
@@ -1,0 +1,6 @@
+export type Course = {
+  id: number;
+  name: string;
+};
+
+export type CourseList = Course[];
